cust: Add empty files for odm bind mounting
The frameworks jars and permissions xmls reside in a ODM partition
folder: since we are forced to get the APKs in /product, we need
to bindmount the ODM jars and xmls there too.

For this reason, copy empty files in place so that we can bind
mount the odm ones there.