cust: Remove CnE stack, moved to device-sony-common and Sony ODM
Clean up the CnE stack as it's now moved to SW binaries
distributed by Sony. Also, the entire setup has been moved
to device-sony-common as to enable the functionality on
Sony devices built as per the Sony AOSP build guide, so
remove it from here.

Only one quirk remains, and that's us re-signing the APKs
for custom ROMs: this is because Sony distributes these
applications signed with Google Android test-keys, while
the vast majority of custom roms are using own keys to
enhance the security of the device.
This also needs "a line or two" to exclude the Sony setup
for the APK bindmount from ODM and vary our configuration
for custom roms needs.